requirements.txt
Django==2.2.12
django-markdownx==3.0.1
django-taggit==1.5.1
gunicorn==20.1.0
Markdown==3.2.1
Pillow==8.3.2
python-dateutil==2.8.1
pytz==2020.1
six==1.16.0
sqlparse==0.3.1
whitenoise==5.3.0
